fix: close the external review's findings across gates, lease, harness and fixture
Thirty-four findings from an external review, fixed in seven groups with
disjoint file sets so the work could run concurrently without merge damage.
The ones worth naming:

A lease handle was made inheritable before the try whose finally clears it, so
anything throwing in between left the desktop lock inheritable for the rest of
the run and every later spawn inherited it - the default-deny leak the design
exists to prevent. The set and the handle it guards now sit inside the try.

Two bounded-spawn paths were not bounded. A failed job-object assignment was
recorded as a note and execution continued, leaving a child with no bound at
all; and a bounded read-drain discarded WaitAll's verdict and read .Result
regardless, so a grandchild holding the pipe write end blocked it forever.

The contract gate could stay green while a rule executed zero checks, because
counts were summed across rules and fixtures bypassed the scan entry point.
Each rule now asserts it ran. Rule 9 approximated "any Native export" with a
name convention that two shipped exports already defeated; it now uses an
explicit table.

The redaction gate scanned two hardcoded roots, so a capture directory added
anywhere else passed silently - the same trap this sub-phase was built to
close, reappearing one layer up. The scan set is derived from tracked files
now, and it went from 23 files to 159. Twenty-one legacy captures from earlier
sub-phases are allowlisted by name rather than silently skipped, because
closing the blind spot should not also mean rewriting files this work never
touched. Its prose limitation is stated rather than papered over: a
regex-shaped prose rule would be exactly as unfalsifiable as a comment saying
read carefully.

The fixture was one 1,191-line file; it is now eleven, each under the cap, as a
pure move verified by comparing the AutomationId set before and after.

Three plan rows named tests that were never shipped. One of them could never
have been shipped: the leg it described reads a property the contract gate
statically bans, so the plan contradicted its own gate. The rows now say what
runs and, where a leg cannot exist, why.

Suite 1174, up from 1169. All gates green.
