+++ docs/plans/2026-08-16-001-feat-windows-2-12-fixture-e2e-harness-plan.md
@@ -782,12 +782,12 @@ Every requirement maps to at least one test that fails if the requirement is vio
 
 | Requirement | Test that fails if violated | Unit |
 |---|---|---|
-| R1 pinned toolchain | `build.ps1`'s invocation asserted to carry `/langversion:5`, plus `csc.exe /langversion:6` on a scratch file exiting non-zero with `CS1617` (the compiler-ceiling proof `A24-1` measured); the optional C# 6 construct leg asserts `CS1056`, the diagnostic this compiler actually emits | U4 |
+| R1 pinned toolchain | `build.ps1` invokes `csc.exe` with `/langversion:5` on every harness run, so fixture source written above the C# 5 ceiling fails that compile with `CS1617` (or `CS1056` for a construct the compiler recognizes but the pinned level rejects) before any scenario runs - the ceiling itself is `A24-1`'s measured fact (the nine-value `/langversion` sweep in `probes/windows/FINDINGS.md`), not a dedicated test; no leg separately re-proves the ceiling by compiling a scratch file at `/langversion:6` | U4 |
 | R2 AutomationId on every target | the precondition leg resolving every inventory id by `--native-id`, aborting and naming the missing id | U5, U9 |
-| R3 independent status readout | the `get --property value` test: value read is the expectation, `--property name` does not contain it, identity unchanged across two `find` calls - which fails against a `Label` | U4 |
+| R3 independent status readout | contract-gate rule 12 forbids every harness file from reading a status through `--property text` or `--property name` (a MUST-CATCH fixture per form), so no leg can read `name` and diff it against the expectation - the discrimination rule 12 itself must supply structurally. `FixtureStatus` (`tests/fixture-app-windows/FixtureStatus.cs`) backs that structurally-forced route with an architecturally two-slot control: `AutomationId`-derived, static `Name` versus counter-suffixed, live `Value`, so a value read can never coincide with a name read | U4, U8 |
 | R4 role vocabulary | the single-snapshot role-coverage leg over R4's closed list, including `cell`; **plus** the Rust test pinning covered ∪ uncovered = the map's producible set | U9, U13 |
 | R5 below-fold ladder | `scroll-to` leg asserting a previously below-fold `scroll-row-N` became visible and `scroll-offset` changed | U9 |
-| R6 non-activating fixture | foreground read via raw `GetForegroundWindow` before and after launch | U4 |
+| R6 non-activating fixture | `Run-E2E.ps1` sets `AGENT_DESKTOP_FIXTURE_NO_ACTIVATE=1` on every launch, and the fixture source shows every window under `ShowWithoutActivation`/`SW_SHOWNOACTIVATE` when that flag is present; no leg re-observes this independently - no raw `GetForegroundWindow` read brackets the fixture's own launch | U4 |
 | R7 independent re-observation | `Assert-Effect` self-tests, executed by U8's `-SelfTest` half against the canned-envelope stub: fails when the status did not change though `ok` was true, **and** fails when the status already equalled the expectation before the action; contract-gate rule 5 with a MUST-CATCH per accessor form | U7, U8 |
 | R8 qualified refs | contract-gate rule 4 with both fixtures; `Find-Target` self-test rejecting a payload without `snapshot_id` | U7, U8 |
 | R9 skips visible | `Write-Verdict` self-tests: undeclared token → non-zero; zero-disposition scenario → non-zero; **all-skipped run → non-zero**; `Add-Fail` → non-zero and named in the verdict | U7, U8 |
@@ -882,7 +882,7 @@ Every requirement maps to at least one test that fails if the requirement is vio
     - §2.14 and §2.15 carry actionable scope text for **all six** deferrals, with §2.15's text naming 2.0's RDP/session-isolation row by its exact `probes/windows/FINDINGS.md` row id **and carrying the runner-registration obligations in full - interactive Task-Scheduler launch, label allocation, trigger re-ratification, fork-PR approval, ephemeral-versus-persistent, credential hygiene, queue flush, and the live gate itself**;
     - **no mention of 2.12 outside §2.12/§2.12.1 claims work this sub-phase routed elsewhere**: §2.4.1's `focused_window` identity sentence and §2.5's HWND wrap-rate sentence are corrected unconditionally, §2.9's `A21-6` parenthetical cites the contended rate measured here, and the CI-requirements row and Risk Register R12 name §2.15 as the registering sub-phase (U1 item 19);
     - and `scripts/check-phases-ledger-citations.ps1` is green - matching case-insensitive substring over the whole file, with a MUST-PASS fixture per stem - so none of the retired phrases above can be reintroduced without a red gate.
-17. Every requirement in the Verification Contract maps to a test that was run; every test added by this sub-phase was invert-verified; and the PR record lists, per live leg, which fixture-side or adapter-side break was used to see it go red.
+17. Every requirement in the Verification Contract maps to the verification its row states. For every requirement but R1's compiler ceiling and R6's non-activating launch, that verification is a test that was run. R1's ceiling is `A24-1`'s measured fact, re-enforced on every run only by the pinned `/langversion:5` compile itself, not re-proven by a dedicated leg; R6's non-activating behavior is a launch configuration (`AGENT_DESKTOP_FIXTURE_NO_ACTIVATE=1`, `ShowWithoutActivation`) with no leg that independently re-observes it. R3's independent-readout guarantee is a test - contract-gate rule 12 and its MUST-CATCH fixtures - that forbids the reading form the row once described rather than one that performs and diffs it. Every test added by this sub-phase was invert-verified; and the PR record lists, per live leg, which fixture-side or adapter-side break was used to see it go red.
 18. All gates green; zero `unwrap()`/`expect()` outside `#[cfg(test)]`; no non-doc comments in `crates/**` or `src/**`; no file over 400 LOC; no delivery-plan references in shipped source; Conventional Commits, authored by Lahfir, no co-authors; PR opened against `feat/windows-adapter`, never `main`.
 
 ---
