+++ scripts/fixtures/capture-redaction/must-catch/path-element-bracket-in-label.json
@@ -1,0 +1,14 @@
+{
+  "probe": "self-test-fixture",
+  "matches": [
+    {
+      "ref_id": "e7",
+      "role": "button",
+      "name": "<redacted:2 chars>",
+      "path": [
+        "window:Report ] Draft",
+        "button:CustomerRealName"
+      ]
+    }
+  ]
+}
