+++ scripts/fixtures/capture-redaction/must-pass/occluder-name-nested-bounds.json
@@ -1,0 +1,17 @@
+{
+  "probe": "self-test-fixture",
+  "error": {
+    "details": {
+      "checks": [
+        {
+          "kind": "occlusion",
+          "occluder": {
+            "role": "window",
+            "name": "<redacted:24 chars>",
+            "bounds": { "x": 100, "y": 200, "width": 300, "height": 150 }
+          }
+        }
+      ]
+    }
+  }
+}
