+++ scripts/fixtures/e2e-windows-contract/must-catch/rule09-add-type-typedefinition.ps1
@@ -1,0 +1,3 @@
+function Invoke-RawPInvokeLeg {
+    Add-Type -Namespace Raw -Name Native -MemberDefinition '[DllImport("user32.dll")] public static extern bool SetForegroundWindow(IntPtr hWnd);'
+}
