+++ scripts/fixtures/e2e-windows-contract/must-catch/rule09-native-e2esecurityattributes.ps1
@@ -1,0 +1,3 @@
+function Invoke-StageLeg {
+    New-E2ESecurityAttributes -Inheritable:$true
+}
