+++ scripts/fixtures/e2e-windows-contract/must-catch/rule09-native-restrictedintegritytoken.ps1
@@ -1,0 +1,3 @@
+function Invoke-StageLeg {
+    New-RestrictedIntegrityToken -IntegrityLevel Low
+}
