+++ scripts/fixtures/e2e-windows-contract/must-catch/rule10-invoke-expression.ps1
@@ -1,0 +1,3 @@
+function Invoke-VersionLeg {
+    Invoke-Expression 'agent-desktop.exe version'
+}
