+++ scripts/fixtures/e2e-windows-contract/must-catch/rule10-process-start.ps1
@@ -1,0 +1,3 @@
+function Invoke-VersionLeg {
+    [System.Diagnostics.Process]::Start('agent-desktop.exe', 'version')
+}
