+++ scripts/fixtures/e2e-windows-contract/must-catch/rule10-start-job.ps1
@@ -1,0 +1,3 @@
+function Invoke-VersionLeg {
+    Start-Job -ScriptBlock { agent-desktop.exe version }
+}
