+++ scripts/fixtures/e2e-windows-contract/must-catch/rule10-variable-assigned.ps1
@@ -1,0 +1,4 @@
+function Invoke-VersionLeg {
+    $binary = 'target\release\agent-desktop.exe'
+    & $binary version
+}
