+++ scripts/fixtures/e2e-windows-contract/must-pass/rule09-add-type-assemblyname.ps1
@@ -1,0 +1,3 @@
+function Invoke-LoadAssemblyLeg {
+    Add-Type -AssemblyName System.Windows.Forms
+}
