+++ scripts/fixtures/e2e-windows-contract/must-pass/rule10-unrelated-variable-call.ps1
@@ -1,0 +1,4 @@
+function Invoke-HelperLeg {
+    $helper = 'Get-Something'
+    & $helper
+}
