+++ scripts/lib/e2e-windows-contract-rules-process.psm1
@@ -114,21 +114,44 @@ function Test-Rule10StagedBinaryInvocation {
     <#
     .SYNOPSIS
         Rule 10: only Invoke-GuardedAgent may invoke the staged
-        agent-desktop.exe. Catches the `&` call operator and Start-Process
+        agent-desktop.exe. Catches the `&` call operator (including on a
+        variable this file itself assigned from a matching string literal -
+        the tracked-assignment half below), Start-Process, Start-Job,
+        Invoke-Expression, and a static [...Process]::Start(...) call, each
         naming something that looks like the staged binary. Scope
         (DesktopLease.psm1/BoundedProcess.psm1/Lib.psm1 implement the
         primitive; selftest/** deliberately drives both guarded paths to
         prove the lease-adoption protocol) is enforced by the caller.
+
+        Residual left deliberately unclosed: a call-operator target bound
+        through a variable whose value cannot be seen as a string literal
+        anywhere in the same file - e.g. read from a parameter, a config
+        file, or built by string concatenation - has no text for this
+        AST-only rule to match against. Only same-file, direct
+        `$var = '...'` assignment is tracked.
     #>
     [CmdletBinding()]
     param($Parsed)
     $hits = New-Object System.Collections.Generic.List[object]
     $bindingRegex = '(?i)(agent-desktop|StagedBinary|AgentDesktopBinary)'
+
+    $assignedBindingVars = New-Object System.Collections.Generic.HashSet[string]
+    foreach ($assign in (Find-E2EAstNodes -Ast $Parsed.Ast -Predicate { $args[0] -is [System.Management.Automation.Language.AssignmentStatementAst] })) {
+        $left = $assign.Left -as [System.Management.Automation.Language.VariableExpressionAst]
+        if (-not $left) { continue }
+        if ($assign.Right.Extent.Text -match $bindingRegex) {
+            [void]$assignedBindingVars.Add($left.VariablePath.UserPath)
+        }
+    }
+
     foreach ($node in (Find-E2EAstNodes -Ast $Parsed.Ast -Predicate { $args[0] -is [System.Management.Automation.Language.CommandAst] })) {
         $ampersand = [System.Management.Automation.Language.TokenKind]::Ampersand
         if ($node.InvocationOperator -eq $ampersand -and $node.CommandElements.Count -gt 0) {
-            $target = $node.CommandElements[0].Extent.Text
-            if ($target -match $bindingRegex) {
+            $targetElement = $node.CommandElements[0]
+            $target = $targetElement.Extent.Text
+            $targetVariable = $targetElement -as [System.Management.Automation.Language.VariableExpressionAst]
+            $matchesTrackedAssignment = $targetVariable -and $assignedBindingVars.Contains($targetVariable.VariablePath.UserPath)
+            if ($target -match $bindingRegex -or $matchesTrackedAssignment) {
                 $hits.Add((New-E2EViolation -RuleId 'rule10' -Pattern 'call-operator-staged-binary' -Line $node.Extent.StartLineNumber -Message "call-operator invocation of the staged binary: $target"))
             }
             continue
@@ -138,8 +161,36 @@ function Test-Rule10StagedBinaryInvocation {
             if ($matched) {
                 $hits.Add((New-E2EViolation -RuleId 'rule10' -Pattern 'start-process-staged-binary' -Line $node.Extent.StartLineNumber -Message 'Start-Process naming the staged binary'))
             }
+            continue
+        }
+        if (Test-E2ECommandName -CommandAst $node -Names @('Start-Job')) {
+            $matched = $node.CommandElements | Where-Object { $_.Extent.Text -match $bindingRegex }
+            if ($matched) {
+                $hits.Add((New-E2EViolation -RuleId 'rule10' -Pattern 'start-job-staged-binary' -Line $node.Extent.StartLineNumber -Message 'Start-Job naming the staged binary'))
+            }
+            continue
+        }
+        if (Test-E2ECommandName -CommandAst $node -Names @('Invoke-Expression', 'iex')) {
+            $matched = $node.CommandElements | Where-Object { $_.Extent.Text -match $bindingRegex }
+            if ($matched) {
+                $hits.Add((New-E2EViolation -RuleId 'rule10' -Pattern 'invoke-expression-staged-binary' -Line $node.Extent.StartLineNumber -Message 'Invoke-Expression naming the staged binary'))
+            }
         }
     }
+
+    foreach ($node in (Find-E2EAstNodes -Ast $Parsed.Ast -Predicate { $args[0] -is [System.Management.Automation.Language.InvokeMemberExpressionAst] })) {
+        if (-not $node.Static) { continue }
+        if ($node.Member.Extent.Text -ine 'Start') { continue }
+        $typeExpr = $node.Expression -as [System.Management.Automation.Language.TypeExpressionAst]
+        if (-not $typeExpr) { continue }
+        $typeName = $typeExpr.TypeName.FullName
+        if ($typeName -inotmatch '(^|\.)Process$') { continue }
+        $matched = $node.Arguments | Where-Object { $_.Extent.Text -match $bindingRegex }
+        if ($matched) {
+            $hits.Add((New-E2EViolation -RuleId 'rule10' -Pattern 'process-start-staged-binary' -Line $node.Extent.StartLineNumber -Message "[$typeName]::Start naming the staged binary"))
+        }
+    }
+
     return $hits.ToArray()
 }
